requirements/3.10.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --output-file=requirements/3.10.txt --resolver=backtracking requirements.in
#
-e file:.
    # via -r requirements.in
alabaster==0.7.12
    # via sphinx
asn1crypto==1.5.1
    # via scramp
async-generator==1.10
    # via
    #   trio
    #   trio-websocket
async-timeout==4.0.2
    # via redis
attrs==22.1.0
    # via
    #   jsonschema
    #   outcome
    #   pytest
    #   trio
azure-core==1.26.1
    # via
    #   azure-storage-blob
    #   msrest
azure-storage-blob==12.14.1
    # via testcontainers
babel==2.11.0
    # via sphinx
bcrypt==4.0.1
    # via paramiko
bleach==5.0.1
    # via readme-renderer
cachetools==5.2.0
    # via google-auth
certifi==2022.9.24
    # via
    #   msrest
    #   requests
    #   selenium
cffi==1.15.1
    # via
    #   cryptography
    #   pynacl
charset-normalizer==2.1.1
    # via requests
clickhouse-driver==0.2.4
    # via testcontainers
codecov==2.1.12
    # via -r requirements.in
commonmark==0.9.1
    # via rich
coverage[toml]==6.5.0
    # via
    #   codecov
    #   pytest-cov
cryptography==36.0.2
    # via
    #   -r requirements.in
    #   azure-storage-blob
    #   paramiko
    #   secretstorage
cx-oracle==8.3.0
    # via testcontainers
deprecation==2.1.0
    # via testcontainers
distro==1.8.0
    # via docker-compose
dnspython==2.2.1
    # via pymongo
docker[ssh]==6.0.1
    # via
    #   docker-compose
    #   testcontainers
docker-compose==1.29.2
    # via testcontainers
dockerpty==0.4.1
    # via docker-compose
docopt==0.6.2
    # via docker-compose
docutils==0.19
    # via
    #   readme-renderer
    #   sphinx
ecdsa==0.18.0
    # via python-jose
entrypoints==0.3
    # via flake8
exceptiongroup==1.0.4
    # via
    #   pytest
    #   trio
flake8==3.7.9
    # via -r requirements.in
google-api-core[grpc]==2.10.2
    # via google-cloud-pubsub
google-auth==2.14.1
    # via google-api-core
google-cloud-pubsub==1.7.2
    # via testcontainers
googleapis-common-protos[grpc]==1.57.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
greenlet==2.0.1
    # via sqlalchemy
grpc-google-iam-v1==0.12.4
    # via google-cloud-pubsub
grpcio==1.50.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.48.2
    # via google-api-core
h11==0.14.0
    # via wsproto
idna==3.4
    # via
    #   requests
    #   trio
imagesize==1.4.1
    # via sphinx
importlib-metadata==5.0.0
    # via
    #   keyring
    #   twine
iniconfig==1.1.1
    # via pytest
isodate==0.6.1
    # via msrest
jaraco-classes==3.2.3
    # via keyring
jeepney==0.8.0
    # via
    #   keyring
    #   secretstorage
jinja2==3.1.2
    # via sphinx
jsonschema==3.2.0
    # via docker-compose
kafka-python==2.0.2
    # via testcontainers
keyring==23.11.0
    # via twine
markupsafe==2.1.1
    # via jinja2
mccabe==0.6.1
    # via flake8
more-itertools==9.0.0
    # via jaraco-classes
msrest==0.7.1
    # via azure-storage-blob
neo4j==5.2.1
    # via testcontainers
oauthlib==3.2.2
    # via requests-oauthlib
outcome==1.2.0
    # via trio
packaging==21.3
    # via
    #   deprecation
    #   docker
    #   pytest
    #   redis
    #   sphinx
paramiko==2.12.0
    # via docker
pg8000==1.29.3
    # via -r requirements.in
pika==1.3.1
    # via testcontainers
pkginfo==1.8.3
    # via twine
pluggy==1.0.0
    # via pytest
protobuf==3.20.3
    # via
    #   google-api-core
    #   google-cloud-pubsub
    #   googleapis-common-protos
    #   grpcio-status
psycopg2-binary==2.9.5
    # via testcontainers
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   python-jose
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pycodestyle==2.5.0
    # via flake8
pycparser==2.21
    # via cffi
pyflakes==2.1.1
    # via flake8
pygments==2.13.0
    # via
    #   readme-renderer
    #   rich
    #   sphinx
pyjwt==2.6.0
    # via python-arango
pymongo==4.3.3
    # via testcontainers
pymssql==2.2.7
    # via testcontainers
pymysql==1.0.2
    # via testcontainers
pynacl==1.5.0
    # via paramiko
pyparsing==3.0.9
    # via packaging
pyrsistent==0.19.2
    # via jsonschema
pysocks==1.7.1
    # via urllib3
pytest==7.2.0
    # via
    #   -r requirements.in
    #   pytest-cov
pytest-cov==4.0.0
    # via -r requirements.in
python-arango==7.5.2
    # via testcontainers
python-dateutil==2.8.2
    # via pg8000
python-dotenv==0.21.0
    # via docker-compose
python-jose==3.3.0
    # via python-keycloak
python-keycloak==2.6.0
    # via testcontainers
pytz==2022.6
    # via
    #   babel
    #   clickhouse-driver
    #   neo4j
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
pyyaml==5.4.1
    # via docker-compose
readme-renderer==37.3
    # via twine
redis==4.3.5
    # via testcontainers
requests==2.28.1
    # via
    #   azure-core
    #   codecov
    #   docker
    #   docker-compose
    #   google-api-core
    #   msrest
    #   python-arango
    #   python-keycloak
    #   requests-oauthlib
    #   requests-toolbelt
    #   sphinx
    #   twine
requests-oauthlib==1.3.1
    # via msrest
requests-toolbelt==0.9.1
    # via
    #   python-arango
    #   python-keycloak
    #   twine
rfc3986==2.0.0
    # via twine
rich==12.6.0
    # via twine
rsa==4.9
    # via
    #   google-auth
    #   python-jose
scramp==1.4.4
    # via pg8000
secretstorage==3.3.3
    # via keyring
selenium==4.6.0
    # via testcontainers
six==1.16.0
    # via
    #   azure-core
    #   bleach
    #   dockerpty
    #   ecdsa
    #   google-auth
    #   grpcio
    #   isodate
    #   jsonschema
    #   paramiko
    #   python-dateutil
    #   websocket-client
sniffio==1.3.0
    # via trio
snowballstemmer==2.2.0
    # via sphinx
sortedcontainers==2.4.0
    # via trio
sphinx==5.3.0
    # via -r requirements.in
sphinxcontrib-applehelp==1.0.2
    # via sphinx
sphinxcontrib-devhelp==1.0.2
    # via sphinx
sphinxcontrib-htmlhelp==2.0.0
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.3
    # via sphinx
sphinxcontrib-serializinghtml==1.1.5
    # via sphinx
texttable==1.6.7
    # via docker-compose
tomli==2.0.1
    # via
    #   coverage
    #   pytest
trio==0.22.0
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.9.2
    # via selenium
twine==4.0.1
    # via -r requirements.in
typing-extensions==4.4.0
    # via azure-core
tzdata==2022.6
    # via pytz-deprecation-shim
tzlocal==4.2
    # via clickhouse-driver
urllib3[socks]==1.26.13
    # via
    #   docker
    #   python-arango
    #   python-keycloak
    #   requests
    #   selenium
    #   twine
webencodings==0.5.1
    # via bleach
websocket-client==0.59.0
    # via
    #   docker
    #   docker-compose
wrapt==1.14.1
    # via testcontainers
wsproto==1.2.0
    # via trio-websocket
zipp==3.10.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
